refactor: consolidate match grades into single config in matching.ts
Replace magic strings and duplicated grade set with a typed `MATCH_GRADES`
array that owns labels, thresholds, and notifiability in one place.
Rename lib/project-match.ts → lib/project-match-notify.ts to reflect purpose.
